DR-13 Move API calls out of post-login handler, use calcExpiryOn to calculate expirty of the access token
