+++ lib/routes/auth/post-login.js
@@ -1,33 +1,6 @@
-const axios = require('axios');
-const { CLIENT_ID, CLIENT_SECRET } = require('../../constants/oauth');
 const { getUserByUsername, updateAccessToken } = require('../../queries/users');
-const { logRequest } = require('../../utils/debug');
-
-function refreshAccessToken(user) {
-  const params = new URLSearchParams({
-    refresh_token: user.refresh_token,
-    grant_type: 'refresh_token',
-  });
-
-  try {
-    return axios.post(
-      'https://www.reddit.com/api/v1/access_token',
-      params,
-      {
-        auth: {
-          username: CLIENT_ID,
-          password: CLIENT_SECRET,
-        },
-        headers: {
-          'Content-Type': 'application/x-www-form-urlencoded',
-        },
-      },
-    );
-  } catch (error) {
-    logRequest(`Error while refreshing access_token ${error.message}`);
-    throw new Error(error.message);
-  }
-}
+const { refreshAccessToken } = require('../../services/reddit');
+const { calcExpiresOn } = require('../../utils/date');
 
 async function postLogin(req, res) {
   const { username } = req.body;
@@ -37,7 +10,7 @@ async function postLogin(req, res) {
   if (!user) {
     const error = new Error('Incorrect combination of username and password');
     error.status = 400;
-    res.status(400).send(error);
+    throw error;
   }
 
   if (new Date(user.expires_on) > new Date()) {
@@ -48,13 +21,14 @@ async function postLogin(req, res) {
     };
   } else {
     const response = await refreshAccessToken(user);
-    const expiresOn = new Date(new Date().getTime() + response.data.expires_in * 1000);
+    const expiresOn = calcExpiresOn(response.data.expires_in);
+
     const updatedToken = await updateAccessToken(response.data.access_token, expiresOn, username);
 
     if (!updatedToken) {
       const error = new Error('Failed to insert refreshed access token into the database');
       error.status = 400;
-      res.status(400).send(error);
+      throw error;
     }
 
     req.session.user = {
