+++ lib/services/reddit.js
@@ -39,7 +39,36 @@ async function getUserInfo(accessToken) {
   }
 }
 
+function refreshAccessToken(user) {
+  const { CLIENT_ID, CLIENT_SECRET } = config.get('oauth');
+
+  const params = new URLSearchParams({
+    refresh_token: user.refresh_token,
+    grant_type: 'refresh_token',
+  });
+
+  try {
+    return axios.post(
+      'https://www.reddit.com/api/v1/access_token',
+      params,
+      {
+        auth: {
+          username: CLIENT_ID,
+          password: CLIENT_SECRET,
+        },
+        headers: {
+          'Content-Type': 'application/x-www-form-urlencoded',
+        },
+      },
+    );
+  } catch (error) {
+    logRequest(`Error while refreshing access_token: ${error.message}`);
+    throw new RedditRequestError(error.message);
+  }
+}
+
 module.exports = {
   exchangeCodeForTokensReq,
   getUserInfo,
+  refreshAccessToken,
 };
